+++ rails-console.sh
@@ -3,6 +3,12 @@
 # This script is part of dora -- Docker container for Rails
 # https://github.com/bovender/dora
 
+if [[ $(id -u -n) != app ]]; then
+  echo "Script was invoked by user '$(id -u -n)'; re-invoking as 'app'..."
+  echo
+  exec setuser app $0
+fi
+
 APP_DIR=/home/app/rails
 
 echo "# dora console helper for Rails"